Seperate components into modules
