+++ flare/components/__init__.py
@@ -1,14 +1,6 @@
-import typing
-
-from flare.components import *
-from flare.context import Context
-from flare.converters import Converter, add_converter
-from flare.internal import install
-from flare.row import Row
-
-__all__: typing.Sequence[str] = ("button", "select", "Context", "Converter", "add_converter", "install", "Row")
-
-__version__ = "0.1.0"
+from .base import *
+from .button import *
+from .select import *
 
 # MIT License
 #
