+++ flare/components/select.py
@@ -1,173 +1,29 @@
 from __future__ import annotations
 
-import abc
 import copy
 import typing as t
 
 import hikari
-import sigparse
 
-from flare.exceptions import ComponentError, MissingRequiredParameterError
-from flare.internal import event_handler, serde
+from flare.components.base import Component
+from flare.exceptions import ComponentError
 
 if t.TYPE_CHECKING:
     from flare import context
 
+__all__: t.Final[t.Sequence[str]] = ("select", "Select")
+
 P = t.ParamSpec("P")
-
-__all__: t.Final[t.Sequence[str]] = ("Component", "button", "Button")
-
-Self = t.TypeVar("Self", bound="Component[...]")
-
-
-class Component(abc.ABC, t.Generic[P]):
-    """
-    An abstract class that all components derive from.
-    """
-
-    def __init__(
-        self,
-        cookie: str | None,
-        callback: t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]],
-    ) -> None:
-        self._custom_id = None
-        self._callback = callback
-        self.cookie = cookie or f"{callback.__name__}.{callback.__module__}"
-
-        self.args = {param.name: param.annotation for param in sigparse.sigparse(callback)[1:]}
-
-        if not self.args:
-            # If no args were passed, calling with_params isn't necessary to construct custom_id
-            self._custom_id = self.cookie
-
-        event_handler.components[self.cookie] = self
-
-    @property
-    def width(self) -> int:
-        """
-        The width of the component.
-        """
-        return 1
-
-    @property
-    def custom_id(self) -> str:
-        """
-        The custom ID of the component.
-        """
-        if self._custom_id is None:
-            raise MissingRequiredParameterError(
-                f"Component received no parameters when it has {len(self.args)}. Did you forget to call `with_params()`?"
-            )
-        return self._custom_id
-
-    @property
-    def callback(
-        self,
-    ) -> t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]]:
-        return self._callback
-
-    def set(self: Self, *_: P.args, **values: P.kwargs) -> Self:
-        new = copy.copy(self)  # Create new instance with params set
-        new._custom_id = serde.serialize(self.cookie, self.args, values)
-        return new
-
-    @abc.abstractmethod
-    def build(self, action_row: hikari.api.ActionRowBuilder) -> None:
-        """Build and append a flare component to a hikari action row."""
-        ...
-
-
-class button:
-    """
-    A button message component.
-
-    Args:
-        label:
-            The label on the button.
-        style:
-            The button style.
-        cookie:
-            An identifier to use for the button. A custom cookie can be supplied so
-            a shorter one is used in serializing and deserializing.
-    """
-
-    def __init__(
-        self,
-        label: str | None,
-        emoji: hikari.Emoji | str | None,
-        style: hikari.ButtonStyle,
-        disabled: bool = False,
-        cookie: str | None = None,
-    ) -> None:
-        self.label = label
-        self.emoji = emoji
-        self.disabled = disabled
-        self.style = style
-        self.cookie = cookie
-
-    def __call__(self, callback: t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]]) -> Button[P]:
-        return Button(
-            callback=callback,
-            label=self.label,
-            emoji=self.emoji,
-            disabled=self.disabled,
-            style=self.style,
-            cookie=self.cookie,
-        )
-
-
-class Button(Component[P]):
-    def __init__(
-        self,
-        *,
-        callback: t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]],
-        label: str | None,
-        emoji: hikari.Emoji | str | None,
-        style: hikari.ButtonStyle,
-        disabled: bool = False,
-        cookie: str | None,
-    ) -> None:
-        super().__init__(cookie, callback)
-        self.label = label
-        self.emoji = emoji
-        self.style = style
-        self.disabled = disabled
-
-        if isinstance(self.emoji, str):
-            self.emoji = hikari.Emoji.parse(self.emoji)
-
-    def build(self, action_row: hikari.api.ActionRowBuilder) -> None:
-        """
-        Build the button into the passed action row.
-        """
-
-        if self.style == hikari.ButtonStyle.LINK:
-            raise ComponentError("Link buttons are not supported.")
-
-        if not self.label and not self.emoji:
-            raise ComponentError("Label and emoji cannot both be empty for button component.")
-
-        button = action_row.add_button(self.style, self.custom_id)
-
-        if self.label:
-            button.set_label(self.label)
-
-        if self.emoji:
-            button.set_emoji(self.emoji)
-
-        button.set_is_disabled(self.disabled)
-
-        button.add_to_container()
 
 
 class select:
     """
-    A select menu message component.
+    Decorator for a select menu message component.
 
     Args:
         options:
             An array of options for the select menu. This must be provided when
-            the class is created or using `SelectMenu.set_options`.
+            the class is created or using `Select.set_options`.
         min_vales:
             The minimum amount of values a user must select.
         max_values:
@@ -197,8 +53,8 @@ class select:
         self.placeholder = placeholder
         self.disabled = disabled
 
-    def __call__(self, callback: t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]]) -> SelectMenu[P]:
-        return SelectMenu(
+    def __call__(self, callback: t.Callable[t.Concatenate[context.Context, P], t.Awaitable[None]]) -> Select[P]:
+        return Select(
             cookie=self.cookie,
             callback=callback,
             options=self.options,
@@ -209,7 +65,7 @@ class select:
         )
 
 
-class SelectMenu(Component[P]):
+class Select(Component[P]):
     def __init__(
         self,
         cookie: str | None,
@@ -229,12 +85,9 @@ class SelectMenu(Component[P]):
 
     @property
     def width(self) -> int:
-        """
-        The width of the component.
-        """
         return 5
 
-    def set(self, *_: P.args, **values: P.kwargs) -> SelectMenu[P]:
+    def set(self, *_: P.args, **values: P.kwargs) -> Select[P]:
         s = super().set(*_, **values)
         # The options array should be different for clones.
         s.options = copy.copy(self.options)
